requirements.txt
aerich==0.8.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncclick==8.1.8.0
asyncpg==0.30.0
click==8.1.8
decouple==0.0.7
dictdiffer==0.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
iso8601==2.1.0
passlib==1.7.4
pydantic==2.11.2
pydantic_core==2.33.1
pypika-tortoise==0.5.0
pytz==2025.2
sniffio==1.3.1
starlette==0.46.1
tomlkit==0.13.2
tortoise-orm==0.24.2
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
